requirements.txt
nsepython
python-telegram-bot==13.15
schedule
